+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## New formats support
 
 - FIGlet Font (FLF)
+- TrueType Collection (TTC)
 
 # Version 0.26.0 (2024-11-07)